packages/core_elements/src/web-animations-next/test/blink/keyframe-properties.html

<!doctype html><title>Web Animations API: Keyframe Property tests</title><script src="testharness/testharness.js"></script><script src="testharness/testharnessreport.js"></script><div id=log></div><div id=div></div><script>test(function(){var a={};Object.defineProperty(a,"width",{value:"200px"});Object.defineProperty(a,"height",{value:"100px",enumerable:true});assert_equals(a.width,"200px","width of keyframe is readable");assert_equals(a.height,"100px","height of keyframe is readable");try{div.animate([a,{height:"200px"}],1)}catch(b){assert_unreached("Mismatched properties - both or neither properties on keyframe were considered.")}},"enumerable keyframe properties tests",{help:"http://dev.w3.org/fxtf/web-animations/#dfn-procedure-for-converting-an-ecmascript-value-to-an-idl-keyframe-object",assert:"Only enumerable properties on keyframes are considered",author:"Shane Stephens"});test(function(){var c=function(){this.width="100px"};c.prototype={height:"100px"};var b=function(){this.top="100px"};b.prototype=Object.create(c.prototype);Object.defineProperty(b.prototype,"left",{value:"100px",enumerable:"true"});var a=new b();try{div.animate([a,{top:"200px",left:"200px",height:"200px"}],1)}catch(d){assert_unreached("Mismatched properties - left, width or height not considered on keyframe.")}},"inherited keyframe properties tests",{help:"http://dev.w3.org/fxtf/web-animations/#dfn-procedure-for-converting-an-ecmascript-value-to-an-idl-keyframe-object",assert:"Only properties in Object.keys on keyframes are considered",author:"Shane Stephens"});</script>
